CapstoneProject/requirements.txt
asgiref==3.7.2
contourpy==1.1.1
crispy-bootstrap5==0.7
cycler==0.12.1
Django==4.2.6
django-admin-interface==0.26.1
django-colorfield==0.10.1
django-crispy-forms==2.1
django-recaptcha==3.0.0
et-xmlfile==1.1.0
fonttools==4.43.1
kiwisolver==1.4.5
matplotlib==3.8.0
mysqlclient==2.2.0
numpy==1.26.1
openpyxl==3.1.2
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
PyMySQL==1.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3.post1
reportlab==4.0.5
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tzdata==2023.3